+++ crates/icp-cli/tests/identity_tests.rs
@@ -1026,7 +1026,7 @@ async fn identity_link_hsm() {
         .arg(&pin_file)
         .assert()
         .success()
-        .stderr(contains("Identity \"hsm-identity\" linked to HSM"));
+        .stderr(contains("Identity `hsm-identity` linked to HSM"));
 
     // Verify the identity appears in the list
     ctx.icp()
